+++ src/idunno.AtProto.Types/CidLink.cs
@@ -3,28 +3,27 @@
 
 using System.Text.Json.Serialization;
 
-namespace idunno.AtProto.Repo;
+namespace idunno.AtProto;
 
 /// <summary>
-/// Contains link information for a blob.
+/// Represents a link to content identified by a <see cref="Cid">content identifier</see>.
 /// </summary>
-[Obsolete("Use " + nameof(AtProto.CidLink) + " instead.")]
-public record BlobReference
+public record CidLink
 {
     /// <summary>
-    /// Creates a new instance of <see cref="BlobReference"/>.
+    /// Creates a new instance of <see cref="CidLink"/>.
     /// </summary>
-    /// <param name="link">The <see cref="Cid">content identifier</see> of the blob.</param>
+    /// <param name="link">The <see cref="Cid">content identifier</see> of the content.</param>
     /// <exception cref="ArgumentNullException">Thrown when <paramref name="link"/> is <see langword="null"/>.</exception>
     [JsonConstructor]
-    public BlobReference(Cid link)
+    public CidLink(Cid link)
     {
         ArgumentNullException.ThrowIfNull(link);
         Link = link;
     }
 
     /// <summary>
-    /// Gets the <see cref="Cid">content identifier</see> of the blob.
+    /// Gets the <see cref="Cid">content identifier</see> of the content.
     /// </summary>
     [JsonInclude]
     [JsonPropertyName("$link")]