+++ src/idunno.Bluesky/Video/BlueskyAgent.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Barry Dorrans. All rights reserved.
+// Copyright (c) Barry Dorrans. All rights reserved.
 // Licensed under the MIT License.
 
 using System.Net;
@@ -259,7 +259,7 @@ public partial class BlueskyAgent
             throw new AuthenticationRequiredException();
         }
 
-        AtProtoHttpResult<Blob> uploadResult = await UploadBlob(
+        AtProtoHttpResult<AtProto.Blob> uploadResult = await UploadBlob(
             captionsAsBytes,
             "text/vtt",
             cancellationToken: cancellationToken).ConfigureAwait(false);